docs: add docs for feature
