+++ tests/integration/conditions_feature/flow.yml
@@ -1,0 +1,25 @@
+!Flow
+version: '1'
+with: {}
+executors:
+- name: first
+- name: exec1
+  uses: ContitionDumpExecutor
+  workspace: ${{ENV.TEMP_WORKSPACE}}
+  uses_metas:
+    name: exec1
+  condition:
+    type: '1'
+- name: exec2
+  needs:
+  - first
+  uses: ContitionDumpExecutor
+  workspace: ${{ENV.TEMP_WORKSPACE}}
+  uses_metas:
+    name: exec2
+  condition:
+    type: '2'
+- name: joiner
+  needs:
+  - exec2
+  - exec1
